js: fix timezone issue
Relative time is used when displaying dates which are less than a day
from now. Like "10 hours ago".

But this relative time is calculated from the time generated by the
mod_autoindex (using the server timezone) and the client current time
(using the client timezone). This two timezones may be different and so
the calculated time is incorrect.

The current time of the server has to be used to fix this behavior.