js: scss: update stretched links #27
On some browsers, the stretched links class doesn't work well inside
a table. The method using the "transform: rotate(0)" does not work
nor does the "position-relative" method on the whole row.

The only thing that works is to use the "position-relative" method on
a single cell. So you have to duplicate it for each cell:
https://github.com/twbs/bootstrap/issues/28608#issuecomment-614881795